Split GL renderer target helpers
