+++ CMakeLists.txt
@@ -256,8 +256,10 @@ target_sources(
         src/gl_renderer.c
         src/gl_renderer_draw_lists.c
         src/gl_renderer_mesh_cache.c
+        src/gl_renderer_postprocess.c
         src/gl_renderer_programs.c
         src/gl_renderer_shaders.c
+        src/gl_renderer_targets.c
         src/gl_renderer_texture_cache.c
         src/image.c
         src/input.c
